pyproject.toml
[project]
name = "kalshi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "connectorx>=0.4.4",
    "polars>=1.33.0",
    "prefect>=3.4.17",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "streamlit>=1.49.1",
    "tqdm>=4.67.1",
]
